requirements.txt
Django~=3.2.10
environs~=9.5.0
django-phonenumber-field==7.0.0